refactor deterministic deployment code
